refactor: move schema generation machinery entirely into schema tool
